requirements.txt
absl-py==0.9.0
albumentations==0.4.6
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cloudpickle==2.0.0
cycler==0.10.0
Cython==0.29.24
decorator==4.4.2
detectron2==0.4+cu101
docopt==0.6.2
easydict==1.9
future==0.18.2
fvcore==0.1.3.post20210317
google-auth==1.20.1
google-auth-oauthlib==0.4.1
grpcio==1.31.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
iopath==0.1.9
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.0
kiwisolver==1.2.0
llvmlite==0.34.0
Markdown==3.2.2
matplotlib==3.3.0
networkx==2.4
numba==0.51.2
numpy==1.19.1
oauthlib==3.1.0
omegaconf==2.0.0
opencv-python==4.4.0.42
opencv-python-headless==4.5.4.58
pandas==1.2.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
portalocker==2.3.2
prompt-toolkit==3.0.5
protobuf==3.12.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydot==1.4.2
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.2.1
Shapely==1.7.1
six==1.15.0
tabulate==0.8.9
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.4
termcolor==1.1.0
tifffile==2020.7.24
torch==1.6.0+cu101
torchvision==0.7.0+cu101
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
yacs==0.1.8
zipp==3.1.0
